chore: extract version related functions to separated file
